start new solver script
